# src/main/java/ar/edu/utn/frc/tup/lc/iv/restTemplate/access/AccessUserAllowedType.java
package ar.edu.utn.frc.tup.lc.iv.restTemplate.access;

import lombok.Data;

/**
 * DTO que representa los tipos de acceso permitidos.
 */
@Data
public class AccessUserAllowedType {

    /**
     * Descripción del tipo de acceso.
     */
    private String description;
}